simplified `onDraw()` in `ScreenAnnotator`
