+++ lib/src/main/java/com/grio/lib/features/editor/RectangleAnnotation.kt
@@ -1,27 +1,37 @@
 package com.grio.lib.features.editor
 
-import android.graphics.Canvas
-import android.graphics.Paint
-import android.graphics.PointF
-import android.graphics.RectF
+import android.graphics.*
+import com.grio.lib.features.editor.ShapeAnnotation.Companion.CLICKABLE_AREA_ALLOWANCE
 import kotlin.math.max
 import kotlin.math.min
 
 data class RectangleAnnotation(
     override var color: String,
     override var size: Float,
+    override var defaultBrush: Paint,
     var start: PointF,
     var end: PointF
 ) : ShapeAnnotation {
 
-    constructor(color: String, x: Float, y: Float): this(color, 15f, PointF(x, y), PointF(x, y))
+    init {
+        defaultBrush.apply {
+            style = Paint.Style.STROKE
+            isAntiAlias = true
+            strokeJoin = Paint.Join.ROUND
+            strokeCap = Paint.Cap.ROUND
+            strokeWidth = size
+            color = Color.parseColor(this@RectangleAnnotation.color)
+        }
+    }
 
-    override fun drawShape(canvas: Canvas?, brush: Paint) {
+    constructor(color: String, x: Float, y: Float) : this(color, 15f, Paint(), PointF(x, y), PointF(x, y))
+
+    override fun drawToCanvas(canvas: Canvas?) {
         canvas?.drawRect(
             min(start.x, end.x),
             min(start.y, end.y),
             max(end.x, start.x),
-            max(end.y, start.y), brush
+            max(end.y, start.y), defaultBrush
         )
     }
 
@@ -31,16 +41,19 @@ data class RectangleAnnotation(
     }
 
     override fun wasSelected(x: Float, y: Float): Boolean {
-        val clickableArea = size * 2
         return when {
             // check left
-            x > start.x - clickableArea && x < start.x + clickableArea && y > start.y - clickableArea && y < end.y + clickableArea -> true
+            x > start.x - CLICKABLE_AREA_ALLOWANCE && x < start.x + CLICKABLE_AREA_ALLOWANCE &&
+                    y > start.y - CLICKABLE_AREA_ALLOWANCE && y < end.y + CLICKABLE_AREA_ALLOWANCE -> true
             // check right
-            x > end.x - clickableArea && x < end.x + clickableArea && y > start.y - clickableArea && y < end.y + clickableArea -> true
+            x > end.x - CLICKABLE_AREA_ALLOWANCE && x < end.x + CLICKABLE_AREA_ALLOWANCE &&
+                    y > start.y - CLICKABLE_AREA_ALLOWANCE && y < end.y + CLICKABLE_AREA_ALLOWANCE -> true
             // check top
-            y > start.y - clickableArea && y < start.y + clickableArea && x > start.x - clickableArea && x < end.x + clickableArea -> true
+            y > start.y - CLICKABLE_AREA_ALLOWANCE && y < start.y + CLICKABLE_AREA_ALLOWANCE &&
+                    x > start.x - CLICKABLE_AREA_ALLOWANCE && x < end.x + CLICKABLE_AREA_ALLOWANCE -> true
             // check bottom
-            y > end.y - clickableArea && y < end.y + clickableArea && x > start.x - clickableArea && x < end.x + clickableArea -> true
+            y > end.y - CLICKABLE_AREA_ALLOWANCE && y < end.y + CLICKABLE_AREA_ALLOWANCE && x
+                    > start.x - CLICKABLE_AREA_ALLOWANCE && x < end.x + CLICKABLE_AREA_ALLOWANCE -> true
             // default
             else -> false
         }
