+++ lib/src/main/java/com/grio/lib/features/editor/ShapeAnnotation.kt
@@ -1,14 +1,13 @@
 package com.grio.lib.features.editor
 
-import android.graphics.Canvas
-import android.graphics.Paint
 import com.grio.lib.features.editor.views.Shape
 
-interface ShapeAnnotation : BugAnnotation {
-    fun drawShape(canvas: Canvas?, brush: Paint)
+interface ShapeAnnotation : BaseAnnotation {
     fun updateShape(x: Float, y: Float)
 
     companion object {
+        const val CLICKABLE_AREA_ALLOWANCE = 40f
+
         fun createShape(color: String, size: Float, x: Float, y: Float, type: Shape): ShapeAnnotation {
             return when (type) {
                 Shape.RECTANGLE -> RectangleAnnotation(color, x, y)
